Move broadcast logic from Clients into Room.
